fix(*): fix todo

+++ test/color_test.cpp
@@ -20,7 +20,7 @@
 
 /// @todo 没有运行
 TEST(color_t, rgba) {
-    color_t rgba(1, 2, 3, 40, color_t::COLOR_ORDER_RGBA);
+    color_t rgba((uint8_t)1, 2, 3, 40, color_t::COLOR_ORDER_RGBA);
     EXPECT_EQ(rgba[0], 1);
     EXPECT_EQ(rgba[1], 2);
     EXPECT_EQ(rgba[2], 3);
@@ -48,7 +48,7 @@ TEST(color_t, rgba) {
 }
 
 TEST(color_t, argb) {
-    color_t argb(40, 1, 2, 3, color_t::COLOR_ORDER_ARGB);
+    color_t argb((uint8_t)40, 1, 2, 3, color_t::COLOR_ORDER_ARGB);
 
     EXPECT_EQ(argb[0], 1);
     EXPECT_EQ(argb[1], 2);
